requirements.txt
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.8.0
scikit-learn==1.3.1
tqdm==4.66.1
pyyaml==6.0.1
